avcodec/dnxhdenc: fix possible out of bound writes for big w/h
It was caused by integer overflows.

+++ libavcodec/dnxhdenc.h
@@ -34,7 +34,7 @@
 #include "dnxhddata.h"
 
 typedef struct RCCMPEntry {
-    uint16_t mb;
+    uint32_t mb;
     int value;
 } RCCMPEntry;
